+++ framework/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './camelCase';
+export * from './deepCloneWithoutUndefined';
 export * from './emptyObject';
 export * from './extractArgumentNames';
 export * from './getEnvVar';