+++ framework/eslint.config.mjs
@@ -17,3 +17,5 @@ export default [
   pluginJs.configs.recommended,
   ...tseslint.configs.recommended
 ];
+
+